made shipping address compulsory to place order, displayed shipping address in related templates

+++ orders/views.py
@@ -9,6 +9,8 @@ from django.contrib import messages
 from user.views import CustomerGroupRequiredMixin
 from django.contrib.auth.models import Group
 from django.db.models import Sum, F, ExpressionWrapper, DecimalField
+from user.forms import AddressRadioForm
+from user.models import Address
 
 class OrderListView(LoginRequiredMixin, ListView):
   model = Order
@@ -79,33 +81,41 @@ class CheckOutView(LoginRequiredMixin, CustomerGroupRequiredMixin, View):
   template_name = 'orders/order_confirmation.html'
 
   def get(self, request, *args, **kwargs):
-    return render(request, self.template_name)
+    # retrieve addresses belonging to the current user
+    addresses = request.user.address_set.all()
+    return render(request, self.template_name, {'addresses': addresses, 'address_form': AddressRadioForm()})
   
   def post(self, request, *args, **kwargs):
     try:
-      if 'confirm_order' in request.POST:
-        order = Order.objects.create(user=request.user)
+      if 'confirm_order' in request.POST:        
+        address_id = request.POST.get('selected_address')
+        if address_id:
+          order = Order.objects.create(user=request.user)
 
-        total_amount = order.calculate_total_amount()
-        order.total_amount = total_amount
+          total_amount = order.calculate_total_amount()
+          order.total_amount = total_amount
+          address = Address.objects.get(pk=address_id)
+          order.shipping_address = address
+          order.save()
 
-        order.save()
-
-        cart_items = request.user.cart.items.all()
-        for cart_item in cart_items:
-          OrderItem.objects.create(
-            order = order,
-            product = cart_item.product,
-            quantity = cart_item.quantity,
-            price = cart_item.price,
-            discount_percentage = cart_item.discount_percentage
-          )
-        # clear user's cart
-        request.user.cart.items.all().delete()
-        # delete the cart
-        request.user.cart.delete()
-        messages.success(request, 'Order added successfully.')
-        return redirect('/orders/')
+          cart_items = request.user.cart.items.all()
+          for cart_item in cart_items:
+            OrderItem.objects.create(
+              order = order,
+              product = cart_item.product,
+              quantity = cart_item.quantity,
+              price = cart_item.price,
+              discount_percentage = cart_item.discount_percentage
+            )
+          # clear user's cart
+          request.user.cart.items.all().delete()
+          # delete the cart
+          request.user.cart.delete()
+          messages.success(request, 'Order added successfully.')
+          return redirect('/orders/')
+        else:
+          messages.error(request, 'Please select a shipping address.')
+          return redirect('checkout')
       else:
         messages.error(request, 'Please confirm your order.')
         return redirect('cart_detail')
